--- src/library/classes/index.ts
@@ -1,5 +1,0 @@
-// -------
-// CLASSES
-// -------
-
-export * from './Parser';
