+++ src/library/index.ts
@@ -3,3 +3,4 @@
 // -------
 
 export * from './CSV';
+export * from './VSCSV';
